increase length of short_message
Can we get at least 50 chr? That's when GitHub says the commit message is going from short to long: "Great commit summaries contain fewer than 50 characters". I'm using this pkg to analyse a git repo and some of key details are the last 30 chrs, so it would be great to get a few more chrs from the parse_log_detailed fn. Thanks!